pyproject.toml
[project]
name = "llm_regression"
version = "0.1.0.dev0"
description = "The llm_regression Python package."
readme = "README.md"
authors = [
  { name="Alex Ioannides", email="alex@bodywork.com" },
]
dependencies = [
  "openai==1.30.*",
  "numpy==1.26.*",
  "pandas==2.2.*",
  "python-dotenv==1.0.*",
  "scikit-learn==1.4.*",
  "tqdm==4.66.*",
]

[project.optional-dependencies]
dev = [
  "black==23.9.1",
  "python-dotenv>=1.0.0",
  "icecream",
  "ipython",
  "mypy==1.5.1",
  "nox==2023.4.22",
  "pandas-stubs==2.2.2.240514",
  "pytest==7.4.2",
  "ruff==0.0.290",
  "types-pytz==2024.1.0.20240417",
  "types-tqdm==4.66.0.20240417",
]
deploy = [
  "build>=1.0.0",
  "pip>=23.2.0",
  "setuptools>=68.0.0",
  "twine>=4.0.0",
  "wheel>=0.41.0",
]

[project.urls]
"Homepage" = "https://github.com/AlexIoannides/llm-regression"
"Bug Tracker" = "https://github.com/AlexIoannides/llm-regression/issues"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.ruff]
src = ["src"]
target-version = "py310"
line-length = 88
select = [
  "D",   # pydocstyle
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
  "W",   # pycodestyle warnings
]
ignore = [
  "D203",    # fix pydocstyle warning
  "D213",    # fix pydocstyle warning
]

[tool.ruff.per-file-ignores]
"tests/*" = [
  "D103",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
files = [
  "src",
  "tests",
  "noxfile.py",
]

[[tool.mypy.overrides]]
module = [
  "sklearn.*",
]
ignore_missing_imports = true
